chore: enable clippy::cast_possible_truncation (#5557)
Description
---
Enables clippy::cast_possible_truncation
Enables clippy::cast_possible_wrap

Motivation and Context
---
This warns us of possible casts in the code base

How Has This Been Tested?
---
unit tests all pass

+++ base_layer/wallet_ffi/wallet.h
@@ -1514,8 +1514,8 @@ TariWalletAddress *liveness_data_get_public_key(TariContactsLivenessData *livene
  * The ```liveness_data_destroy``` method must be called when finished with a TariContactsLivenessData to prevent a
  * memory leak
  */
-int liveness_data_get_latency(TariContactsLivenessData *liveness_data,
-                              int *error_out);
+unsigned int liveness_data_get_latency(TariContactsLivenessData *liveness_data,
+                                       int *error_out);
 
 /**
  * Gets the last_seen time (in local time) from a TariContactsLivenessData
@@ -3133,8 +3133,8 @@ unsigned long long wallet_get_fee_estimate(struct TariWallet *wallet,
                                            unsigned long long amount,
                                            struct TariVector *commitments,
                                            unsigned long long fee_per_gram,
-                                           unsigned long long num_kernels,
-                                           unsigned long long num_outputs,
+                                           unsigned int num_kernels,
+                                           unsigned int num_outputs,
                                            int *error_out);
 
 /**
